refactor: implement UnorderedList styled component
